CrateDB: Use a random integer instead of autoincremented `experiment_id`
The original database schema, defined by SQLAlchemy models, uses
`autoincrement=True` for the `experiment_id` column. Because CrateDB
does not provide such a feature, this swaps in a different approach to
generate a random and unique integer value.

For this to make work, the database column data type had to be updated
to `BIGINT` instead of `INTEGER`.